+++ app/components/primer/beta/spinner.html.erb
