+++ docs/src/@primer/gatsby-theme-doctocat/nav.yml
@@ -88,7 +88,7 @@
   - title: SegmentedControl
     url: "/components/alpha/segmentedcontrol"
   - title: Spinner
-    url: "/components/spinner"
+    url: "/components/beta/spinner"
   - title: State
     url: "/components/state"
   - title: Subhead
