+++ static/arguments.json
@@ -1958,6 +1958,32 @@
     ]
   },
   {
+    "component": "Spinner",
+    "status": "beta",
+    "source": "https://github.com/primer/view_components/tree/main/app/components/primer/beta/spinner.rb",
+    "lookbook": "https://primer.style/view-components/lookbook/inspect/primer/beta/spinner_preview/default/",
+    "parameters": [
+      {
+        "name": "size",
+        "type": "Symbol",
+        "default": "`:medium`",
+        "description": "One of `[:large, 64]`, `[:medium, 32]`, or `[:small, 16]`."
+      },
+      {
+        "name": "style",
+        "type": "String",
+        "default": "`box-sizing: content-box; color: var(--color-icon-primary);`",
+        "description": "Custom element styles."
+      },
+      {
+        "name": "system_arguments",
+        "type": "Hash",
+        "default": "N/A",
+        "description": "[System arguments](/system-arguments)"
+      }
+    ]
+  },
+  {
     "component": "Text",
     "status": "beta",
     "source": "https://github.com/primer/view_components/tree/main/app/components/primer/beta/text.rb",
@@ -2348,32 +2374,6 @@
         "type": "Array<Hash>",
         "default": "`[]`",
         "description": "List of icons to render, in the format { symbol: :icon_name, size: :small }"
-      }
-    ]
-  },
-  {
-    "component": "Spinner",
-    "status": "beta",
-    "source": "https://github.com/primer/view_components/tree/main/app/components/primer/spinner_component.rb",
-    "lookbook": "https://primer.style/view-components/lookbook/inspect/primer/spinner_preview/default/",
-    "parameters": [
-      {
-        "name": "size",
-        "type": "Symbol",
-        "default": "`:medium`",
-        "description": "One of `[:large, 64]`, `[:medium, 32]`, or `[:small, 16]`."
-      },
-      {
-        "name": "style",
-        "type": "String",
-        "default": "`box-sizing: content-box; color: var(--color-icon-primary);`",
-        "description": "Custom element styles."
-      },
-      {
-        "name": "system_arguments",
-        "type": "Hash",
-        "default": "N/A",
-        "description": "[System arguments](/system-arguments)"
       }
     ]
   },
