+++ static/statuses.json
@@ -60,6 +60,7 @@
   "Primer::Beta::Popover": "beta",
   "Primer::Beta::ProgressBar": "beta",
   "Primer::Beta::RelativeTime": "beta",
+  "Primer::Beta::Spinner": "beta",
   "Primer::Beta::Text": "beta",
   "Primer::Beta::Truncate": "beta",
   "Primer::Beta::Truncate::TruncateText": "alpha",
@@ -86,7 +87,7 @@
   "Primer::OcticonComponent": "deprecated",
   "Primer::OcticonSymbolsComponent": "alpha",
   "Primer::PopoverComponent": "deprecated",
-  "Primer::SpinnerComponent": "beta",
+  "Primer::SpinnerComponent": "deprecated",
   "Primer::StateComponent": "beta",
   "Primer::SubheadComponent": "beta",
   "Primer::TabContainerComponent": "alpha",
